Скорректирована логика приложения под работу unit-тестов
* при корректировке логики использовал достаточно "грязный" подход исправления кода без рефакторинга

+++ src/main/java/com/game/controller/PlayersController.java
@@ -11,6 +11,7 @@ import java.util.function.Predicate;
 import java.util.stream.Collectors;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
+import org.springframework.lang.NonNull;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -329,7 +330,11 @@ public class PlayersController {
    */
   @GetMapping(value = "/{playerId}")
   public ResponseEntity<PlayerDto> getPlayer(@PathVariable long playerId) {
-    System.out.println(">> getPlayer(): playerId = " + playerId);
+
+    if (playerId <= 0) {
+      throw new IllegalArgumentException("ID can't be zero");
+    }
+
     PlayerDto player = playerService.getPlayer(playerId);
     return ResponseEntity.ok(player);
   }
@@ -342,7 +347,19 @@ public class PlayersController {
   @PostMapping(value = "/{playerId}")
   public ResponseEntity<PlayerDto> changePlayer(
       @PathVariable long playerId, @RequestBody PlayerDto player) {
+
+    if (playerId <= 0) {
+      throw new IllegalArgumentException("ID can't be zero");
+    }
+
+    if (player.isEmpty()) {
+      return ResponseEntity.ok(playerService.getPlayer(playerId));
+    }
+
     player.setId(playerId);
+
+    validateUpdateDto(player);
+
     PlayerDto updatedPlayer = playerService.changePlayer(player);
     return ResponseEntity.ok(updatedPlayer);
   }
@@ -357,8 +374,37 @@ public class PlayersController {
   }
 
   private void validateDto(PlayerDto player) {
-    if (player.getBirthday() <= 0) {
+
+    if (player.isEmpty()) {
+      throw new IllegalArgumentException("PlayerDTO can't be empty");
+    }
+    if (player.getBirthday() < 94667400000L || player.getBirthday() > 3250366919900L) {
       throw new IllegalArgumentException("Birthday can't be less zero");
+    }
+    if (player.getTitle().length() > 30) {
+      throw new IllegalArgumentException("Title is too big");
+    }
+    if (player.getName().length() > 12) {
+      throw new IllegalArgumentException("Name is too big");
+    }
+    if (player.getExperience() < 0 || player.getExperience() > 10_000_000) {
+      throw new IllegalArgumentException("Bad experience");
+    }
+  }
+
+  private void validateUpdateDto(PlayerDto player) {
+
+    if (player.getBirthday() != null && (player.getBirthday() < 94667400000L || player.getBirthday() > 3250366919900L)) {
+      throw new IllegalArgumentException("Birthday can't be less zero");
+    }
+//    if (player.getTitle().length() > 30) {
+//      throw new IllegalArgumentException("Title is too big");
+//    }
+//    if (player.getName().length() > 12) {
+//      throw new IllegalArgumentException("Name is too big");
+//    }
+    if (player.getExperience() != null && (player.getExperience() < 0 || player.getExperience() > 10_000_000)) {
+      throw new IllegalArgumentException("Bad experience");
     }
   }
 
@@ -370,6 +416,11 @@ public class PlayersController {
    */
   @DeleteMapping(value = "{playerId}")
   public ResponseEntity<String> deletePlayer(@PathVariable long playerId) {
+
+    if (playerId <= 0) {
+      throw new IllegalArgumentException("ID can't be zero");
+    }
+
     playerService.deletePlayer(playerId);
     return ResponseEntity.ok(String.format("Player with id %d deleted!", playerId));
   }
